pyproject.toml
[tool.poetry]
name = "bqat"
version = "0.1.0"
description = ""
authors = ["L.Wu <l.wu@biometix.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
click = "^8.1.3"
rich = "^13.3.1"
pendulum = "^2.1.2"
docker = "^6.0.1"
pydantic = "^1.10.5"


[tool.poetry.group.dev]
optional = true


[tool.poetry.group.dev.dependencies]
tox = "^4.4.6"
pytest = "^7.2.2"
black = "^23.1.0"
mypy = "^1.0.1"
coverage = "^7.2.1"
genbadge = {extras = ["coverage", "tests"], version = "^1.1.0"}
pytest-html = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
bqat = 'bqat.app:run'